requirements.txt
allure-pytest==2.13.5
pytest==8.3.3
requests==2.32.3
jsonschema==4.23.0
pytest-xdist==3.6.1